adding routing module, setting up url path parameters for the vehicle-add component
