+++ clientApp/src/app/checkposts/checkposts.component.css
@@ -2,7 +2,11 @@ h2 {
   margin-top: 5%; }
 
 .checkpost-panel {
-  margin: 0 0 10%; }
+  cursor: pointer; }
+  .checkpost-panel > .row {
+    margin: 0 0 10%; }
+    .checkpost-panel > .row:hover {
+      box-shadow: 0 0 24px 0 rgba(0, 0, 0, 0.35); }
   .checkpost-panel .fa-tasks {
     margin-top: 4px; }
   .checkpost-panel .fa-tasks, .checkpost-panel p {
@@ -23,8 +27,8 @@ h2 {
   border: #f5f5f5 solid 1px; }
 
 .checkpost-comparison-top {
-  background-color: #f5f5f5;
   padding-top: 12px;
-  padding-bottom: 12px; }
+  padding-bottom: 12px;
+  background-color: #f5f5f5; }
 
-/*# sourceMappingURL=app.component.css.map */
+/*# sourceMappingURL=checkposts.component.css.map */
