+++ clientApp/src/app/checkposts/checkposts.component.scss
@@ -4,8 +4,14 @@ $lightgray: #f5f5f5;
 h2 {
 	margin-top: 5%;
 }
-.district-panel {
-	margin: 0 0 10%;
+.checkpost-panel {
+	cursor: pointer;
+	&>.row {
+		margin: 0 0 10%;
+		&:hover {
+			box-shadow: 0 0 24px 0 rgba(0,0,0,.35);
+		}
+	}
 	.fa-tasks {
 		margin-top: 4px;
 	}
@@ -14,23 +20,23 @@ h2 {
 	}
 	p {
 		font-size: 2rem;
-		&.district-quantity {
+		&.checkpost-quantity {
 			font-size: 1rem;
 		}
 	}
 }
-.district-panel-top {
+.checkpost-panel-top {
 	background-color: $blue;
 }
-.district-panel-bottom {
+.checkpost-panel-bottom {
 	padding: 8px;
 	border: $blue 1px solid;
 }
-.district-comparison {
+.checkpost-comparison {
 	border: $lightgray solid 1px;
 }
-.district-comparison-top {
-	background-color: $lightgray;
+.checkpost-comparison-top {
 	padding-top: 12px;
 	padding-bottom: 12px;
+	background-color: $lightgray;
 }
